Fix LLP64 (Windows) 32-bit shift bugs in motif matching
On LLP64 platforms (Windows) `long`/`int` are 32-bit, so several bit masks
built by shifting a 32-bit literal by up to 60-62 positions were undefined
behaviour and produced corrupted masks:

- IUPACMotif::includes last-cell mask (1L -> (cell)1): enumeratePatterns on
  long, mostly-N patterns matched ~96% of sequences instead of the true
  frequency (e.g. spaced dyads reported in nearly every promoter).
- Pattern k-mer bitset stored in `unsigned long` (32-bit) while indexing bits
  0-63 (MASK_SHIFT=6 / MASK_FILTER=63): bits 32-63 aliased onto 0-31, causing
  false-positive cores in the spaced-dyad / repeat counters.
- Reverse-complement / last-cell masks in the motif builders, incl. a
  shift-by-cellSize UB in fillComplementBuf.

Fixed by performing the shifts in 64-bit ((cell)1 / (uint64_t)1 / ~(cell)0).
After the fix, forward-strand enumeratePatterns and enumerateDyadsWithCore
agree exactly with an independent regex count on real promoters. Adds
regression tests covering pattern lengths that straddle the 16-nt cell
boundary. A residual reverse-complement bug for >16 nt non-palindromic motifs
is documented in NEWS.md (its regression test is skipped).

+++ src/Pattern/Pattern.cpp
@@ -69,18 +69,18 @@ bool Pattern::check(unsigned kmer) const {
 	if ((kmer >> MASK_SHIFT) >= maskSize) {
 		return false;
 	}
-	return mask[kmer >> MASK_SHIFT] & (((unsigned long) 1) << (kmer &  MASK_FILTER));
+	return mask[kmer >> MASK_SHIFT] & (((uint64_t) 1) << (kmer &  MASK_FILTER));
 }
 
 void Pattern::markPattern(unsigned kmer) {
 	if ((kmer >> MASK_SHIFT) < maskSize) {
-		mask[kmer >> MASK_SHIFT] |= ((unsigned long) 1) << (kmer &  MASK_FILTER);
+		mask[kmer >> MASK_SHIFT] |= ((uint64_t) 1) << (kmer &  MASK_FILTER);
 	}
 }
 
 void Pattern::init(std::string pattern) {
 	patternSize = pattern.length();
-	maskSize = 1 << (2*patternSize - MASK_SHIFT);
+	maskSize = ((uint64_t) 1) << (2*patternSize - MASK_SHIFT);
 
 	mask.resize(maskSize, 0);
 	add(pattern);
